pyproject.toml
[project]
name = "trail-cam-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.56.0",
    "numpy",
    "opencv-python",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
